requirements.txt
absl-py==2.1.0
adtk==0.6.2
aioitertools
aiosignal
akracer==0.0.13
akshare==1.12.20
alabaster
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
alpaca-py==0.18.1
altair==5.2.0
anaconda-project
appdirs==1.4.4
applaunchservices
appnope
appscript
APScheduler==3.10.4
archspec
argon2-cffi
argon2-cffi-bindings
arrow
astroid
asttokens
astunparse==1.6.3
atomicwrites==1.4.0
attrs==25.4.0
Automat
autopep8
Babel
backports.functools-lru-cache
backports.tempfile
backports.weakref==1.0.post1
backtrader==1.9.78.123
baidu-aip==4.16.13
baostock==0.8.8
bcrypt
binaryornot
bleach
blis==0.7.11
boltons
bottle==0.13.4
Brotli
bs4==0.0.1
build==1.3.0
cachetools==5.3.2
caerus==0.1.10
catalogue==2.0.10
certifi==2026.1.4
chardet
charset-normalizer
clickhouse-connect==0.9.2
cloudpathlib==0.16.0
clr-loader==0.2.6
clyent==1.2.2
colorama
comtypes==1.3.1
conda-libmamba-solver
conda-token
conda-verify==3.4.2
conda_index
confection==0.1.4
crcmod==1.7
cssutils==2.9.0
cycler
cymem==2.0.8
Cython==0.29.37
czsc==0.10.3
dashscope==1.14.0
dataframe-image==0.2.3
decorator
defusedxml
Deprecated==1.2.18
diff-match-patch
dill
diskcache==5.6.3
dm-tree==0.1.8
docstring-to-markdown
docutils
easytrader==0.23.0
easyutils==0.1.7
entrypoints
erl==0.0.0
et-xmlfile==1.1.0
executing
ez_setup==0.9
fake-useragent==1.4.0
fastai==2.7.14
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema
fastprogress==1.0.3
findpeaks==2.6.4
flatbuffers==24.3.7
fonttools==4.25.0
frozendict==2.4.0
func-timeout==4.3.5
future
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.41
gmpy2
google-pasta==0.2.0
GPy==1.13.1
GPyOpt==1.2.6
grpcio==1.62.1
h11==0.16.0
h5py==3.10.0
HeapDict
html2image==2.0.4.3
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperlink
idna
imagesize
incremental
inflection==0.5.1
iniconfig
intake
intervaltree
ipython-genutils
isort
itemadapter
itsdangerous
jaraco.classes
jedi
jieba==0.42.1
jmespath==0.10.0
json5
jsonpath==0.82.2
jsonpointer==2.1
jupyterlab-pygments
keras==3.0.5
kiwisolver
langcodes==3.3.0
lark-oapi==1.4.23
lazy-object-proxy
Levenshtein==0.23.0
libarchive-c
libclang==16.0.6
lightweight-charts==2.1
linkify-it-py
locket
loguru==0.7.3
Markdown
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
meta==1.0.2
mistune
ml-dtypes==0.3.2
mpl-finance==0.10.1
mplfinance==0.12.10b0
msgpack
multipledispatch
multiprocess
multitasking==0.0.11
munkres==1.1.4
murmurhash==1.0.10
MyTT==2.9.3
namex==0.0.7
numpydoc
openpyxl
opt-einsum==3.3.0
oss2==2.19.1
outcome==1.3.0.post0
pandas==2.2.2
pandas_ta==0.3.14b0
pandocfilters
panel
paramz==0.9.6
parse==1.20.2
parso
pathlib
pathspec
patsy==0.5.3
pdfkit==1.0.0
peewee==3.17.1
pep8==1.7.1
pexpect
pickleshare
piecewise-regression==1.5.0
plotly-express==0.4.1
pluggy
ply==3.11
preshed==3.0.9
pretty-html-table==0.9.16
prettytable==3.9.0
progressbar==2.5
progressbar2==4.3.2
prometheus-client
Protego
protobuf==3.20.3
proxy_tools==0.1.0
psutil
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess
pure-eval
pwlf==2.2.1
py-mini-racer==0.6.0
pyarrow==21.0.0
pyasn1
pyasn1-modules==0.2.8
pycparser
pycryptodome==3.23.0
pyct
pydantic_core==2.16.3
pydash==7.0.6
pydeck==0.8.1b0
PyDispatcher==2.0.5
pydocstyle
pyDOE==0.3.8
pyecharts==2.0.4
pyerfa
PyExecJS==1.5.1
pygame==2.5.2
PyJWT
pylint
pyls-spyder==0.4.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyobjc-framework-Security==11.1
pyobjc-framework-WebKit==11.1
pyparsing
pypinyin==0.50.0
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pyquery==2.0.0
PySocks
pytdx
python-dateutil
python-dotenv
python-Levenshtein==0.23.0
python-lsp-jsonrpc
python-retry==0.0.1
python-slugify
python-snappy
python-utils==3.8.1
pythonnet==3.0.3
pyts==0.13.0
pywebview==6.0
pywencai==0.12.2
pywinauto==0.6.6
QDarkStyle
qq-botpy==1.2.0
qqbot==2.3.11
qstock==1.3.6
qstylizer
QtAwesome
rapidfuzz==3.6.1
redis==6.4.0
redmail==0.6.0
requests-file
responses
retrying==1.3.4
rs_czsc==0.1.16.post251006
Rtree
ruamel-yaml-conda
ruamel.yaml
seaborn
selenium==4.39.0
semver
service-identity
simplejson==3.19.2
six
smart-open
smartmoneyconcepts==0.0.24
smmap==5.0.1
snowballstemmer
snownlp==0.12.3
sortedcontainers
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
SQLAlchemy-Utils==0.41.1
srsly==2.4.8
sseclient-py==1.8.0
stack-data
stock-indicators==1.3.1
stockstats==0.6.2
streamlit==1.30.0
stumpy==1.13.0
swig==4.2.1
TA-Lib==0.6.7
tblib
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado
text-unidecode
textdistance
thefuzz==0.19.0
thinc==8.2.3
threadpoolctl
three-merge
tinycss2
tldextract
toml
tomli
tomlkit
toolz
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
traitlets
trio==0.32.0
trio-websocket==0.12.2
tsai==0.3.9
tushare==1.3.6
typer==0.9.0
typing_extensions==4.10.0
tzdata
tzlocal==5.2
uc-micro-py
ujson
Unidecode
urllib3==2.6.3
validators==0.22.0
vnpy-ifind==1.0.2
wasabi==1.1.2
watchdog
wcwidth
weasel==0.3.4
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==11.0.3
wget==3.2
whatthepatch
wrapt
wrds==3.2.0
wsproto==1.3.2
wurlitzer
xlrd==2.0.1
xlwings
xxhash
xyzservices
yfinance==0.2.37
zigzag==0.3.2
zope.interface
zstandard